+++ QModel/src/models/experiment.py
@@ -1,0 +1,1 @@
+# Use curve fit to extrqapolate saved bounds and distances to an extrapolated curve.