+++ QModel/src/models/q_multi_model.py
@@ -271,8 +271,7 @@ class QMultiModel:
         Example:
             best_hyperparams = self.tune(evaluations=300)
         """
-        print(
-            f"[STATUS] Running model tuning for {evaluations} max iterations")
+        print(f"[STATUS] Running model tuning for {evaluations} max iterations")
         space = {
             "max_depth": hp.choice("max_depth", np.arange(1, 20, 1, dtype=int)),
             "eta": hp.uniform("eta", 0, 1),
@@ -427,8 +426,7 @@ class QPredictor:
 
         adj_prediction = prediction * adjustment
         lq_idx = next((i for i, x in enumerate(adj) if x == 1), -1) + i
-        uq_idx = next((i for i, x in reversed(
-            list(enumerate(adj))) if x == 1), -1) + i
+        uq_idx = next((i for i, x in reversed(list(enumerate(adj))) if x == 1), -1) + i
         return adj_prediction, (lq_idx, uq_idx)
 
     def find_and_sort_peaks(self, signal):
@@ -459,7 +457,7 @@ class QPredictor:
         slope_change = []
         for i in range(len(diff) - window_size + 1):
             # Calculate the slope change over the current window
-            window_slope_change = np.mean(np.diff(diff[i: i + window_size]))
+            window_slope_change = np.mean(np.diff(diff[i : i + window_size]))
             slope_change.append(window_slope_change)
 
         slope_change = np.array(slope_change)
@@ -522,8 +520,7 @@ class QPredictor:
 
     def adjustment_poi_1(self, guess, diss_raw, actual):
 
-        zero_slope = self.find_zero_slope_regions(
-            self.normalize(diss_raw), 0.0075, 100)
+        zero_slope = self.find_zero_slope_regions(self.normalize(diss_raw), 0.0075, 100)
         adjusted_guess = guess
 
         if len(zero_slope) >= 2:
@@ -686,8 +683,7 @@ class QPredictor:
             closest_rf_point = filtered_rf_points[closest_rf_idx]
 
             # Calculate distances from candidate points to the closest RF point
-            distances_to_closest_rf = np.abs(
-                candidate_points - closest_rf_point)
+            distances_to_closest_rf = np.abs(candidate_points - closest_rf_point)
 
             # Find the closest candidate point to the closest RF point
             closest_candidate_idx = np.argmin(distances_to_closest_rf)
@@ -784,8 +780,7 @@ class QPredictor:
             closest_rf_point = filtered_rf_points[closest_rf_idx]
 
             # Calculate distances from candidate points to the closest RF point
-            distances_to_closest_rf = np.abs(
-                candidate_points - closest_rf_point)
+            distances_to_closest_rf = np.abs(candidate_points - closest_rf_point)
 
             # Find the closest candidate point to the closest RF point
             closest_candidate_idx = np.argmin(distances_to_closest_rf)
@@ -816,71 +811,16 @@ class QPredictor:
         else:
             return guess
 
-    def adjustment_poi_6(self, guess, diff, actual, poi_5_guess):
-        adjustment = guess
+    def adjustment_poi_6(self, guess, signal, diff, diss, actual, poi_5_guess):
+
+        combo = self.normalize(signal[poi_5_guess:]) + self.normalize(
+            diff[poi_5_guess:]
+        )
+        adjustment = np.argmax(combo) + poi_5_guess
 
         def nearest_peak(peaks, point):
             nearest_peak_idx = np.argmin(np.abs(peaks - point))
             return peaks[nearest_peak_idx]
-
-        def decomp(data, zero_slope_threshold=5e-2):
-            # Compute the first derivative (difference between consecutive points)
-            derivatives = np.diff(data)
-
-            # Initialize lists to store the start and end indices of each region
-            increasing_regions = []
-            decreasing_regions = []
-            zero_slope_regions = []
-
-            # Initialize variables to track the current region type and its start index
-            current_region = None
-            start_idx = 0
-
-            # Iterate through the derivative to find regions
-            for i, slope in enumerate(derivatives):
-                if slope > zero_slope_threshold:
-                    if current_region != 'increasing':
-                        if current_region is not None:
-                            if current_region == 'increasing':
-                                increasing_regions.append((start_idx, i))
-                            elif current_region == 'decreasing':
-                                decreasing_regions.append((start_idx, i))
-                            elif current_region == 'zero':
-                                zero_slope_regions.append((start_idx, i))
-                        current_region = 'increasing'
-                        start_idx = i
-                elif slope < -zero_slope_threshold:
-                    if current_region != 'decreasing':
-                        if current_region is not None:
-                            if current_region == 'increasing':
-                                increasing_regions.append((start_idx, i))
-                            elif current_region == 'decreasing':
-                                decreasing_regions.append((start_idx, i))
-                            elif current_region == 'zero':
-                                zero_slope_regions.append((start_idx, i))
-                        current_region = 'decreasing'
-                        start_idx = i
-                else:
-                    if current_region != 'zero':
-                        if current_region is not None:
-                            if current_region == 'increasing':
-                                increasing_regions.append((start_idx, i))
-                            elif current_region == 'decreasing':
-                                decreasing_regions.append((start_idx, i))
-                            elif current_region == 'zero':
-                                zero_slope_regions.append((start_idx, i))
-                        current_region = 'zero'
-                        start_idx = i
-
-            # Append the last region
-            if current_region == 'increasing':
-                increasing_regions.append((start_idx, len(data) - 1))
-            elif current_region == 'decreasing':
-                decreasing_regions.append((start_idx, len(data) - 1))
-            elif current_region == 'zero':
-                zero_slope_regions.append((start_idx, len(data) - 1))
-
-            return increasing_regions, decreasing_regions, zero_slope_regions
 
         def envelope(signal):
             data = np.array(signal)
@@ -891,64 +831,16 @@ class QPredictor:
 
             return upper_envelope, maxima_indices
 
-        ue, maxima_indices = envelope(diff)
-        increasing, decreasing, zero = decomp(diff)
-        slope = np.diff(diff)
-        ue_slope = np.diff(np.diff(np.diff(ue)))
-        slope_change = np.diff(slope)
-        slope_change_inv = slope_change * -1
-        diff_peaks, _ = find_peaks(diff)
-        slope_peaks_1, _ = find_peaks(slope_change)
-        slope_peaks_2, _ = find_peaks(slope_change_inv)
-        slope_peaks = np.concatenate((slope_peaks_1, slope_peaks_2))
-        ue_peaks, _ = find_peaks(ue_slope)
-        diff = self.normalize(diff)
-        if len(increasing) < 1:
-            return guess
-        divider = decreasing[-1][0]
-        if len(slope_peaks[slope_peaks > divider]) < 1:
-            return guess
-        adjustment = nearest_peak(slope_peaks[slope_peaks > divider], divider)
-        x = np.arange(len(diff))
-        if abs(adjustment - actual[5]) > 300:
-            fig, ax = plt.subplots()
-            ax.plot(self.normalize(ue))
-            ax.plot(self.normalize(ue_slope), color='tan')
-            ax.scatter(maxima_indices, diff[maxima_indices], marker="x")
-            ax.scatter(slope_peaks, diff[slope_peaks])
-            ax.scatter(diff_peaks, diff[diff_peaks], marker='v')
-            ax.scatter(ue_peaks, diff[ue_peaks], marker='*')
-            ax.plot(self.normalize(slope_change), label="2nd Deriv")
-
-            ax.plot(self.normalize(diff), label="diff", color="grey")
-            ax.axvline(guess, color="green", linestyle="dotted",
-                       label="Initial guess")
-            ax.axvline(actual[5], color="orange",
-                       linestyle="--", label="Actual")
-            # Plot filled regions for increasing sections
-            ax.axvline(divider, linestyle='dotted',
-                       color='black', label='divider')
-            # for (l, r) in increasing:
-            #     plt.fill_between(x[l:r+2], diff[l:r+2], color='green',
-            #                      alpha=0.5, label="Increasing" if l == increasing[0][0] else "")
-
-            # # Plot filled regions for decreasing sections
-            # for (l, r) in decreasing:
-            #     plt.fill_between(x[l:r+2], diff[l:r+2], color='red',
-            #                      alpha=0.5, label="Decreasing" if l == decreasing[0][0] else "")
-
-            # # Plot filled regions for zero slope sections
-            # for (l, r) in zero:
-            #     plt.fill_between(x[l:r+2], diff[l:r+2], color='blue',
-            #                      alpha=0.5, label="Zero Slope" if l == zero[0][0] else "")
-
-            # ax.axvline(poi_5_guess, label="POI_5 Guess")
-            # ax.axvline(move_1, color="pink", label="Move 1")
-            # ax.axvline(move_2, color="purple", label="Move 2")
-            ax.axvline(adjustment, label="Adjustment", color="red")
-            plt.legend()
-            plt.show()
-        return guess
+        # if abs(adjustment - actual[5]) > 300:
+        #     fig, ax = plt.subplots()
+        #     ax.plot(self.normalize(diff), label="diff", color="brown")
+        #     ax.plot(self.normalize(diss), label="diss", color="grey")
+        #     ax.plot(combo, label="combo", color="black")
+        #     ax.axvline(adjustment, label="Adjustment", color="red")
+        #     ax.axvline(actual[5], label="Actual", color="red", linestyle="--")
+        #     plt.legend()
+        #     plt.show()
+        return adjustment
 
     def predict(self, file_buffer, type=-1, start=-1, stop=-1, act=None):
         # Load CSV data and drop unnecessary columns
@@ -1097,8 +989,7 @@ class QPredictor:
             poi_1_guess=poi_1,
         )
         # poi_2 = emp_points[1]
-        poi_4 = self.adjustmet_poi_4(
-            df, candidates_4, extracted_4, act[3], bounds_4)
+        poi_4 = self.adjustmet_poi_4(df, candidates_4, extracted_4, act[3], bounds_4)
         poi_5 = self.adjustmet_poi_5(
             df, candidates_5, extracted_5, emp_points[4], act[4], bounds_5
         )
@@ -1107,7 +998,9 @@ class QPredictor:
         poi_3 = np.argmax(adj_3)
         poi_6 = self.adjustment_poi_6(
             np.argmax(adj_6),
+            adj_6,
             df["Difference"],
+            df["Dissipation"],
             act,
             poi_5,
         )
@@ -1119,8 +1012,14 @@ class QPredictor:
             poi_5,
             poi_6,
         ]
+        candidates_1 = np.insert(candidates_1, 0, poi_1)
+        candidates_2 = np.insert(candidates_2, 0, poi_2)
+        candidates_3 = np.insert(candidates_3, 0, poi_3)
+        candidates_4 = np.insert(candidates_4, 0, poi_4)
+        candidates_5 = np.insert(candidates_5, 0, poi_5)
+        candidates_6 = np.insert(candidates_6, 0, poi_6)
         candidates = [
-            [poi_1, adj_1],
+            candidates_1,
             candidates_2,
             candidates_3,
             candidates_4,